refactor: remove repository interface from book entity
