+++ src/books/domain/books/book.go
@@ -1,26 +1,9 @@
-package models
+package books
 
 import (
-	"errors"
 	"gorm.io/gorm"
 	"time"
 )
-
-var ErrProductNotFound = errors.New("product not found")
-
-type BookRepositoryReader interface {
-	GetById(ID uint) (*Book, error)
-	GetAll() ([]*Book, error)
-}
-
-type BookRepositoryWriter interface {
-	Create(book *Book) (*Book, error)
-}
-
-type BookRepository interface {
-	BookRepositoryReader
-	BookRepositoryWriter
-}
 
 // TODO: Separate gorm model from domain model
